+++ packages/types/src/files.ts
@@ -38,6 +38,16 @@ export type ListFilesResponse = {
   cursor?: string;
 };
 
+export type AddFileArguments = {
+  organizationID?: string;
+  userID: string;
+  name: string;
+  description?: string;
+  rolesRequired: string[];
+  mimeType: string;
+  file: File;
+};
+
 export type CreateFileArguments = {
   organizationID?: string;
   blobID: string;
